Fix https://github.com/KillerInstinct/pylnker/pull/4
Note: Small breaking change on how we handle the `Data` item in the `Data_After_EOF`. Instead of assuming it's text, it is now base64 encoded since the input could also be data (or garbage).

Don't bother trying to encode the data just to get it to pass JSON serialization. Historic samples had text (.bat/.ps1 stuff) but you can put anything there... So just Bas64 encode it. That way if you want to eventually write the data out to some file somewhere you can just decode it and do it easily.